requirements.txt
azure-functions
psycopg2-binary
azure-identity
azure-keyvault-secrets
sqlalchemy
requests
pandas
numpy
pgvector
heyoo
openai
scikit-learn
googlemaps